ai_agent/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.4